LocalTempDBCleanup/requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bidict==0.21.2
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
filelock==3.0.12
gitdb==4.0.5
GitPython==3.1.8
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mysql-connector==2.2.9
mysql-connector-python==8.0.25
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
plyer==2.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.17.3
PyAutoIt==0.6.0
pycparser==2.20
Pygments==2.7.1
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
schedule==0.6.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
urllib3==1.26.2
virtualenv==20.0.31
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
